+++ tests/primitives/never_ty/never_ty.c
@@ -20,6 +20,8 @@ int32_t never_ty_never_ty_eta2__id(void const *env_, int32_t arg_);
 
 void never_ty_never_ty_opaqueStmt(void);
 
+_Noreturn void MiloneCore_Prelude_exit(int32_t);
+
 void never_ty_never_ty_graceful(void);
 
 _Noreturn void never_ty_never_ty_crash(void);
@@ -111,7 +113,8 @@ void never_ty_never_ty_graceful(void) {
             goto clause_3;
     }
 clause_2:;
-    exit(0);
+    MiloneCore_Prelude_exit(0);
+    milone_never_error("never_ty/never_ty.milone", 9, 13);
 clause_3:;
     switch_ = 0;
     goto switch_next_1;
@@ -121,17 +124,18 @@ switch_next_1:;
 
 _Noreturn void never_ty_never_ty_crash(void) {
     never_ty_never_ty_opaqueStmt();
-    exit(1);
+    MiloneCore_Prelude_exit(1);
+    milone_never_error("never_ty/never_ty.milone", 13, 0);
 }
 
 void never_ty_never_ty_testStmt(void) {
     never_ty_never_ty_crash();
-    exit(1);
+    milone_never_error("never_ty/never_ty.milone", 19, 7);
 }
 
 void never_ty_never_ty_testLetInit(void) {
     never_ty_never_ty_crash();
-    exit(1);
+    milone_never_error("never_ty/never_ty.milone", 24, 2);
 }
 
 void never_ty_never_ty_testMatchArm(void) {
@@ -147,7 +151,7 @@ void never_ty_never_ty_testMatchArm(void) {
     }
 clause_5:;
     never_ty_never_ty_crash();
-    exit(1);
+    milone_never_error("never_ty/never_ty.milone", 30, 16);
 clause_6:;
     switch_1 = 1;
     goto switch_next_4;
@@ -162,7 +166,7 @@ _Noreturn void never_ty_never_ty_eta2_testFunObj_crash(void const *env_1) {
 
 _Noreturn void never_ty_never_ty_testFunObj_callFnObj(struct UnitNeverFun1 crashFn_) {
     crashFn_.fun(crashFn_.env);
-    exit(1);
+    milone_never_error("never_ty/never_ty.milone", 38, 2);
 }
 
 void never_ty_never_ty_testFunObj(void) {
@@ -191,7 +195,7 @@ void never_ty_never_ty_testOwn(void) {
     }
 clause_8:;
     never_ty_never_ty_crash();
-    exit(1);
+    milone_never_error("never_ty/never_ty.milone", 49, 11);
 clause_9:;
     switch_2 = 0;
     goto switch_next_7;
@@ -202,7 +206,7 @@ switch_next_7:;
 
 _Noreturn void never_ty_never_ty_testInfiniteLoop_diverge(void) {
     never_ty_never_ty_testInfiniteLoop_diverge();
-    exit(1);
+    milone_never_error("never_ty/never_ty.milone", 57, 2);
 }
 
 void never_ty_never_ty_testInfiniteLoop(void) {
@@ -225,13 +229,13 @@ void never_ty_never_ty_testNeverAsValue(void) {
     opt_ = variant_;
     if ((opt_.discriminant != Neveroption1_Some)) goto next_11;
     n_ = opt_.Neveroption1_Some;
-    exit(1);
+    milone_never_error("never_ty/never_ty.milone", 65, 15);
 next_11:;
     if ((opt_.discriminant != Neveroption1_None)) goto next_12;
     never_ty_never_ty_crash();
-    exit(1);
+    milone_never_error("never_ty/never_ty.milone", 66, 18);
 next_12:;
-    exit(1);
+    milone_exhaust_error("never_ty/never_ty.milone", 64, 3);
 end_match_10:;
     variant_1 = (struct Int32NeverResult2_){.discriminant = Int32NeverResult2_Ok, .Int32NeverResult2_Ok = 0};
     ok_ = variant_1;
@@ -242,9 +246,9 @@ end_match_10:;
 next_14:;
     if ((ok_.discriminant != Int32NeverResult2_Error)) goto next_15;
     n_1 = ok_.Int32NeverResult2_Error;
-    exit(1);
+    milone_never_error("never_ty/never_ty.milone", 73, 17);
 next_15:;
-    exit(1);
+    milone_exhaust_error("never_ty/never_ty.milone", 71, 4);
 end_match_13:;
     value_ = match_1;
     if ((value_ != 0)) milone_assert_error("never_ty/never_ty.milone", 75, 2);
@@ -263,7 +267,7 @@ int main(int argc, char **argv) {
     never_ty_never_ty_testInfiniteLoop();
     never_ty_never_ty_testNeverAsValue();
     never_ty_never_ty_crash();
-    exit(1);
+    milone_never_error("never_ty/never_ty.milone", 78, 0);
 }
 
 void never_ty_never_ty_start(void) {